port amplitude_corrected_multioctave_noise, and fix the persistence operand's width (#226)
`elevation_lakes` and `elevation_nauvis` both read this wrapper, so phase 6
cannot skip it. `fixtures.rs` used to say it was "deliberately NOT ported yet:
porting it now would mean porting a known-wrong model and enshrining its
wrongness in a Rust assertion" - that line is now the opposite of the port's
standing rule, which is to reproduce the TypeScript faithfully so tier 2 stays
honest and to fix a model in a change graded on its own.

It lands at 81/152 with worst 1.788139e-7, both measured on the TypeScript
side against the same fixture. The count is FROZEN rather than bounded, which
is strictly stronger than what the TypeScript asserts (`worst < 2.5e-7`, no
count): a change that moved 30 positions inside that bound would pass there
and fail here.

## The port scored BETTER than the TypeScript, and finding out why moved a
## shipped op

The first working version scored 89/152 at worst 5.960e-8. The cause was not
the wrapper: `variable_persistence_multioctave_noise` took `persistence` as an
**f32** and the TypeScript multiplies its f32 accumulator by an un-narrowed
JavaScript number - `f32(acc * persistence)`.

No fixture in the tree could see that. `oracle-variable-persistence-multioctave`'s
captured `persistenceField` is `0.35 + 0.25 * basis_noise{...}` evaluated in
the noise machine, so all 38 values are exactly f32 and both widths score
266/266 with worst 0. It is the same shape as #191 and #309 - a narrowing the
fixtures agree on because they only ever offer values already on the grid.

`oracle-multioctave-wrappers`'s amplitude-corrected cases DO discriminate,
because they pass the program constant `0.7` straight in and 0.7 is not
f32-exact:

| persistence operand           | exact   | worst     |
| ----------------------------- | ------- | --------- |
| f64 (the TypeScript, and now this) | 81/152 | 1.788e-7 |
| f32 (what shipped here)            | 89/152 | 5.960e-8 |

The better-scoring form is NOT the one taken. 89 is an improvement and not a
full exact count, so the greedy-accept rule rejects it as a model change, and
adopting it would put a divergence into every Nauvis elevation value with
nothing to grade it. It is a real finding for #254 all the same: that issue
records the 81/152 as unexplained, and this names one term in it, worth 8
points and still 63 short. Neither form is the game's.

## Two harness compensations removed

- `checksum_variable_persistence` crossed the ABI as an f32, so
  `test/wasmMultioctaveParity.spec.ts` narrowed its own value with
  `Math.fround` before comparing - making the two sides agree by construction
  on exactly the term that differed. Both now use f64. Two of that spec's
  cases (0.62, 0.9) are not f32-exact, so the comparison grades the operand
  width instead of hiding it: planting the f32 narrowing back into the Rust
  turns it red at `p=0.62`, checked rather than claimed.
- `p ** octaves` is `powf`, not `powi`. `powi` disagrees with V8 by one ULP at
  0.7^4, 0.7^6 and 0.7^8, and one ULP there flips the f32 rounding of the
  octave gain, which moves every point in the case.
  `the_geometric_series_matches_v8s_pow_bit_for_bit` pins the raw bits at all
  nine (persistence, octaves) pairs this port can reach.

`engine.wasm` 178,531 -> 178,533 bytes.

+++ crates/fmw-wasm/src/lib.rs
@@ -153,6 +153,14 @@ pub extern "C" fn checksum_multioctave_noise(
 /// where a difference would read as an op divergence. The per-tile path is
 /// graded by tier 1 instead, which feeds the fixture's captured
 /// `persistenceField`; the spec calls this with several values.
+///
+/// It crosses as an **f64**, and that is load-bearing rather than tidy. It was
+/// an f32 until #226, so the spec narrowed its own value with `Math.fround`
+/// before comparing - a harness compensation that made the two sides agree by
+/// construction on exactly the term that turned out to differ. Two of that
+/// spec's cases pass a persistence which is not f32-exact (0.62 and 0.9), so
+/// with the compensation gone this comparison now grades the operand width
+/// instead of hiding it.
 #[unsafe(no_mangle)]
 #[allow(clippy::too_many_arguments)]
 pub extern "C" fn checksum_variable_persistence(
@@ -162,7 +170,7 @@ pub extern "C" fn checksum_variable_persistence(
     input_scale: f64,
     output_scale: f64,
     offset_x: f64,
-    persistence: f32,
+    persistence: f64,
     x0: f64,
     y0: f64,
     step: f64,
